rename module bapsflib._hdf.maps.controls.clparse -> bapsflib._hdf.maps.controls.parsers
