--- docs/api_static/bapsflib._hdf.maps.controls.clparse.rst
@@ -1,9 +1,0 @@
-bapsflib\.\_hdf\.maps\.controls\.clparse
-========================================
-
-.. py:currentmodule:: bapsflib._hdf.maps.controls.clparse
-
-.. automodapi:: bapsflib._hdf.maps.controls.clparse
-    :no-heading:
-    :include-all-objects:
-    :headings: "-^"
