+++ docs/api_static/bapsflib._hdf.maps.controls.parsers.rst
@@ -1,0 +1,9 @@
+bapsflib\.\_hdf\.maps\.controls\.parsers
+========================================
+
+.. py:currentmodule:: bapsflib._hdf.maps.controls.parsers
+
+.. automodapi:: bapsflib._hdf.maps.controls.parsers
+    :no-heading:
+    :include-all-objects:
+    :headings: "-^"
